Make DataCheckerPermissionPacket an inner class of DataCheckerPacket
