+++ src/main/java/com/mraof/minestuck/network/DataCheckerPacket.java
@@ -2,6 +2,7 @@ package com.mraof.minestuck.network;
 
 import com.mraof.minestuck.Minestuck;
 import com.mraof.minestuck.client.gui.playerStats.DataCheckerScreen;
+import com.mraof.minestuck.player.ClientPlayerData;
 import com.mraof.minestuck.skaianet.DataCheckerManager;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.FriendlyByteBuf;
@@ -76,6 +77,36 @@ public final class DataCheckerPacket
 		{
 			if(packetIndex == DataCheckerPacket.index)
 				DataCheckerScreen.activeComponent = new DataCheckerScreen.MainComponent(nbtData);
+		}
+	}
+	
+	public record Permission(boolean isAvailable) implements MSPacket.PlayToClient
+	{
+		public static final ResourceLocation ID = Minestuck.id("data_checker/permission");
+		
+		@Override
+		public ResourceLocation id()
+		{
+			return ID;
+		}
+		
+		@Override
+		public void write(FriendlyByteBuf buffer)
+		{
+			buffer.writeBoolean(isAvailable);
+		}
+		
+		public static Permission read(FriendlyByteBuf buffer)
+		{
+			boolean isAvailable = buffer.readBoolean();
+			
+			return new Permission(isAvailable);
+		}
+		
+		@Override
+		public void execute()
+		{
+			ClientPlayerData.handleDataPacket(this);
 		}
 	}
 }
