+++ src/main/java/com/mraof/minestuck/util/DataCheckerPermission.java
@@ -4,7 +4,7 @@ import com.mraof.minestuck.Minestuck;
 import com.mraof.minestuck.MinestuckConfig;
 import com.mraof.minestuck.computer.editmode.EditData;
 import com.mraof.minestuck.computer.editmode.ServerEditHandler;
-import com.mraof.minestuck.network.data.DataCheckerPermissionPacket;
+import com.mraof.minestuck.network.DataCheckerPacket;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.players.ServerOpListEntry;
@@ -62,12 +62,12 @@ public class DataCheckerPermission
 	
 	private static void sendPacket(ServerPlayer player)
 	{
-		DataCheckerPermissionPacket packet;
+		DataCheckerPacket.Permission packet;
 		boolean permission = hasPermission(player);
 		if(permission)
 			dataCheckerPermission.add(player.getGameProfile().getId());
 		else dataCheckerPermission.remove(player.getGameProfile().getId());
-		packet = new DataCheckerPermissionPacket(permission);
+		packet = new DataCheckerPacket.Permission(permission);
 		PacketDistributor.PLAYER.with(player).send(packet);
 	}
